% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg.R
\docType{data}
\name{ecg}
\alias{ecg}
\title{5 minutes of yours truely ECG}
\format{A numeric vector of 60000 points in uV.}
\source{
\url{http://fullpower.com}
}
\usage{
ecg
}
\description{
The data set is a reading from a Philips Alice Sleep Diagnostic System
}
\keyword{datasets}

